requirements.txt
# Core dependencies for ig2rss

# Instagram API client
instagrapi>=2.1.2

# HTTP requests (for media downloads)
requests>=2.31.0

# HTTP server
flask>=3.0.3

# Job scheduling
APScheduler>=3.10.4

# Environment variable management
python-dotenv>=1.0.1
